Workarena/Base: Merge init scripts before evaluation.
Fixes #57 and duplicates without the need for downgrading the playwright version.

Playwright's init scripts are not executed sequentially, as per PW's docs:

<img width="979" height="175" alt="image" src="https://github.com/user-attachments/assets/1d9a9c9a-150f-40de-b4f7-6def953d7dc9" />

However, some benchmarks depend on all installed init scripts to be executed in order, which is why the page hangs if they don't (in WorkArena's case, this was because of the function being waited on being undefined - since the script defining it hadn't been evaluated yet).

It _shouldn't_ cause any problem to just concatenate all scripts together, so that we can guarantee FIFO execution order.